refactor: split oversized harness modules
