+++ app/renderer/sections/Advisor.tsx
@@ -5,33 +5,21 @@ import type { DateRange, MenubarPayload, Period } from '../lib/types'
 import type { MetroraHarnessActionEvent } from '../lib/metrora-bridge-types'
 import { createAdvisorDataSource } from '../advisor/source'
 import { createAdvisorKernel } from '../advisor/kernel'
-import { createAdvisorRuntime } from '../advisor/runtime'
 import { createAdvisorOverviewSnapshot } from '../advisor/tools'
-import { HostedAdvisorRuntime, probeHostedAdvisor } from '../advisor/hosted'
 import { advisorContextualSurfaceLabel, advisorScopeFromContextualLaunch, normalizeAdvisorContextualLaunch, type AdvisorContextualLaunchV1, type AdvisorContextualScopeMode } from '../advisor/context'
 import { advisorScopeForRequestedPeriod } from '../advisor/turn-plan'
-import { advisorScopeFingerprint, type AdvisorAnswer, type AdvisorConversationTurn, type AdvisorHostedProviderId, type AdvisorLocalRuntimeId, type AdvisorReasoningEffort, type AdvisorScope, type AdvisorScopeConflictOptionV1, type AdvisorScopeConflictV1 } from '../advisor/types'
+import { advisorScopeFingerprint, type AdvisorAnswer, type AdvisorConversationTurn, type AdvisorScope, type AdvisorScopeConflictOptionV1, type AdvisorScopeConflictV1 } from '../advisor/types'
 import { contextualScopeLabel } from './advisor-scope-labels'
-import { createHostedProbeChecking, createHostedProbeFailure, presentHostedProbe, type HarnessHostedProbePresentation, type HarnessRuntimeChoice } from '../harness/HarnessRuntimePopover'
-import { AdvisorHostedOperationGuard, isSelectableHostedModel } from './advisor-hosted-operation-guard'
 import { harnessToolLabel, type HarnessToolActivity } from '../harness/HarnessWorkTrace'
 import { HarnessSurface } from '../harness/HarnessSurface'
 import { isSwarmExperimentalEnabled } from '../swarm/feature-gate'
 import { useSwarmRun } from '../swarm/useSwarmRun'
-import { isAdvisorCancelled, useAdvisorLocalRuntime } from './useAdvisorLocalRuntime'
+import { isAdvisorCancelled } from './useAdvisorLocalRuntime'
+import { useAdvisorRuntimeController } from './useAdvisorRuntimeController'
 type DetectedProvider = { id: string; label: string }
 type AdvisorMessage = { id: string; role: 'user' | 'assistant'; text?: string; answer?: AdvisorAnswer; scopeFingerprint: string }
 type AdvisorConversation = { id: string; title: string; messages: AdvisorMessage[] }
 type AdvisorFailedRequest = { question: string; scope: AdvisorScope; conversationId: string; conversation: AdvisorConversationTurn[] }
-const DEFAULT_REASONING_EFFORTS: readonly AdvisorReasoningEffort[] = ['default']
-const REASONING_EFFORT_STORAGE_KEY = 'metrora.harness.reasoning-effort'
-
-function storedReasoningEffort(): AdvisorReasoningEffort {
-  try {
-    const value = window.localStorage.getItem(REASONING_EFFORT_STORAGE_KEY)
-    return value === 'low' || value === 'medium' || value === 'high' || value === 'max' || value === 'default' ? value : 'default'
-  } catch { return 'default' }
-}
 
 function makeId(prefix: string): string {
   return prefix + '-' + Date.now().toString(36) + '-' + Math.random().toString(36).slice(2, 7)
@@ -87,54 +75,6 @@ export function Advisor({
     if (scope.model && !modelOptions.includes(scope.model)) setScope(current => ({ ...current, model: null }))
   }, [modelOptions, scope.model])
   const source = useMemo(() => createAdvisorDataSource(metrora), [])
-  const fallbackRuntime = useMemo(() => createAdvisorRuntime(), [])
-  const [runtimeChoice, setRuntimeChoice] = useState<HarnessRuntimeChoice>('ollama')
-  const [hostedProvider, setHostedProvider] = useState<AdvisorHostedProviderId>('openai')
-  const hostedOperationGuardRef = useRef(new AdvisorHostedOperationGuard(hostedProvider))
-  const [hostedModel, setHostedModel] = useState<string | null>(null)
-  const hostedModelRef = useRef<string | null>(null)
-  hostedModelRef.current = hostedModel
-  const [hostedConsent, setHostedConsent] = useState(false)
-  const [reasoningEffort, setReasoningEffort] = useState<AdvisorReasoningEffort>(storedReasoningEffort)
-  const [hostedProbe, setHostedProbe] = useState<HarnessHostedProbePresentation>(() => createHostedProbeChecking('openai'))
-  const hostedModelForRuntime = hostedModel ? hostedProbe.models.find(model => model.id === hostedModel && isSelectableHostedModel(model)) ?? null : null
-  const hostedRuntime = useMemo(() => hostedModelForRuntime ? new HostedAdvisorRuntime({ provider: hostedProvider, model: hostedModelForRuntime.id, capabilities: hostedModelForRuntime.capabilities, reasoningEffort, consent: hostedConsent }) : null, [hostedConsent, hostedModelForRuntime, hostedProvider, reasoningEffort])
-  const markHostedModelVerified = useCallback((provider: AdvisorHostedProviderId, model: string) => {
-    if (!hostedOperationGuardRef.current.isCurrentProvider(provider) || hostedModelRef.current !== model) return
-    setHostedProbe(current => current.provider !== provider ? current : {
-      ...current,
-      models: current.models.map(item => item.id === model && item.state !== 'unsupported' && item.state !== 'failed-conformance'
-        ? {
-            ...item,
-            state: 'verified',
-            limitation: 'This exact model passed a bounded Metrora Harness request; deterministic evidence retrieval remains authoritative.',
-          }
-        : item),
-    })
-  }, [])
-  const [configureOpen, setConfigureOpen] = useState(false)
-  const { runtimeId, setRuntimeId, runtimeModel, setRuntimeModel, runtimeState, llamaServerPort, localRuntime, checkLocalRuntime, setLocalModel, setLlamaServerPort } = useAdvisorLocalRuntime()
-  const activeRuntime = runtimeChoice === 'hosted'
-    ? hostedRuntime ?? fallbackRuntime
-    : localRuntime ?? fallbackRuntime
-  const reasoningEfforts = activeRuntime.reasoningEfforts ?? DEFAULT_REASONING_EFFORTS
-  const effectiveReasoningEffort = reasoningEfforts.includes(reasoningEffort) ? reasoningEffort : 'default'
-  useEffect(() => {
-    try { window.localStorage.setItem(REASONING_EFFORT_STORAGE_KEY, reasoningEffort) } catch { /* unavailable in restricted contexts */ }
-  }, [reasoningEffort])
-  const kernel = useMemo(() => createAdvisorKernel(source, activeRuntime), [activeRuntime, source])
-  const swarmExperimentalEnabled = isSwarmExperimentalEnabled()
-  const [mode, setMode] = useState<'chat' | 'swarm'>('chat')
-  const [swarmConversationId, setSwarmConversationId] = useState<string | null>(null)
-  const swarm = useSwarmRun({
-    source,
-    runtime: activeRuntime,
-    scope,
-    overview: !overview.loading && !overview.switching && scope.period === period && scope.provider === provider && scope.projectId === projectScopeId && scope.range?.from === range?.from && scope.range?.to === range?.to && scope.model === null && overview.data ? createAdvisorOverviewSnapshot(scope, overview.data) : null,
-    modelId: runtimeChoice === 'hosted' ? hostedModel ?? activeRuntime.id : runtimeModel ?? activeRuntime.id,
-    modelLabel: runtimeChoice === 'hosted' ? hostedModelForRuntime?.label ?? activeRuntime.label : runtimeModel ?? activeRuntime.label,
-    enabled: swarmExperimentalEnabled,
-  })
   const [loadingQuestion, setLoadingQuestion] = useState<string | null>(null)
   const [streamPreview, setStreamPreview] = useState('')
   const [toolStatus, setToolStatus] = useState<string | null>(null)
@@ -150,65 +90,40 @@ export function Advisor({
     setToolStatus(null)
     setToolActivity([])
   }, [])
-  const hostedProbeController = useRef<AbortController | null>(null)
-  const checkHostedRuntime = useCallback(async (requestedProvider: AdvisorHostedProviderId = hostedProvider, resetSelection = false) => {
-    if (!hostedOperationGuardRef.current.isCurrentProvider(requestedProvider)) return
-    invalidateAdvisorRequest()
-    hostedProbeController.current?.abort()
-    const requestId = hostedOperationGuardRef.current.startProbe(requestedProvider)
-    if (requestId === null) return
-    const controller = new AbortController()
-    hostedProbeController.current = controller
-    const isCurrentRequest = () => !controller.signal.aborted && hostedOperationGuardRef.current.isCurrentProbe(requestedProvider, requestId)
-    setHostedProbe(current => createHostedProbeChecking(requestedProvider, current))
-    try {
-      const result = await probeHostedAdvisor(requestedProvider, controller.signal)
-      if (!isCurrentRequest()) return
-      setHostedProbe(presentHostedProbe(result))
-      const selectable = result.models.find(isSelectableHostedModel)
-      if (result.available && selectable) {
-        const currentModel = resetSelection ? null : hostedModelRef.current
-        const next = currentModel && result.models.some(model => model.id === currentModel && isSelectableHostedModel(model)) ? currentModel : selectable.id
-        setHostedModel(next)
-        if (next !== currentModel) setHostedConsent(false)
-      } else {
-        setHostedModel(null)
-        setHostedConsent(false)
-      }
-    } catch (caught) {
-      if (isCurrentRequest() && !isAdvisorCancelled(caught)) {
-        setHostedModel(null)
-        setHostedConsent(false)
-        setHostedProbe(createHostedProbeFailure(requestedProvider))
-      }
-    }
-  }, [hostedProvider, invalidateAdvisorRequest])
-  useEffect(() => {
-    if (runtimeChoice !== 'hosted') return
-    void checkHostedRuntime()
-    return () => hostedProbeController.current?.abort()
-  }, [checkHostedRuntime, runtimeChoice])
+  const [notice, setNotice] = useState<string | null>(null)
+  const {
+    activeRuntime,
+    runtimeChoice,
+    hostedProvider,
+    hostedModel,
+    hostedModelForRuntime,
+    hostedConfigRef,
+    hostedSubmitBlockReason,
+    runtimeModel,
+    runtimeState,
+    checkLocalRuntime,
+    markHostedModelVerified,
+    runtimeControls,
+  } = useAdvisorRuntimeController({ invalidateAdvisorRequest, setNotice })
+  const kernel = useMemo(() => createAdvisorKernel(source, activeRuntime), [activeRuntime, source])
+  const swarmExperimentalEnabled = isSwarmExperimentalEnabled()
+  const [mode, setMode] = useState<'chat' | 'swarm'>('chat')
+  const [swarmConversationId, setSwarmConversationId] = useState<string | null>(null)
+  const swarm = useSwarmRun({
+    source,
+    runtime: activeRuntime,
+    scope,
+    overview: !overview.loading && !overview.switching && scope.period === period && scope.provider === provider && scope.projectId === projectScopeId && scope.range?.from === range?.from && scope.range?.to === range?.to && scope.model === null && overview.data ? createAdvisorOverviewSnapshot(scope, overview.data) : null,
+    modelId: runtimeChoice === 'hosted' ? hostedModel ?? activeRuntime.id : runtimeModel ?? activeRuntime.id,
+    modelLabel: runtimeChoice === 'hosted' ? hostedModelForRuntime?.label ?? activeRuntime.label : runtimeModel ?? activeRuntime.label,
+    enabled: swarmExperimentalEnabled,
+  })
   const [conversations, setConversations] = useState<AdvisorConversation[]>(() => [{ id: makeId('chat'), title: 'New chat', messages: [] }])
   const [activeConversationId, setActiveConversationId] = useState(() => conversations[0]!.id)
   const [historyQuery, setHistoryQuery] = useState('')
   const [composer, setComposer] = useState('')
-  const [credentialEntry, setCredentialEntry] = useState('')
-  const [credentialSaving, setCredentialSaving] = useState(false)
-  const hostedConfigRef = useRef({ runtimeChoice, hostedRuntime, hostedConsent })
-  hostedConfigRef.current = { runtimeChoice, hostedRuntime, hostedConsent }
-  const hostedSubmitBlockReason = runtimeChoice === 'hosted'
-    ? !hostedRuntime
-      ? hostedProbe.reachability === 'checking'
-        ? 'Waiting for the hosted provider check to finish.'
-        : 'Connect a hosted provider credential and select a usable model before sending a message.'
-      : !hostedConsent
-        ? 'Confirm the hosted-provider prompt and evidence sharing notice before sending.'
-        : null
-    : null
-  useEffect(() => { setCredentialEntry('') }, [hostedProvider])
   const [error, setError] = useState<string | null>(null)
   const [failedRequest, setFailedRequest] = useState<AdvisorFailedRequest | null>(null)
-  const [notice, setNotice] = useState<string | null>(null)
   const [selectedAnswerId, setSelectedAnswerId] = useState<string | null>(null)
   const [harnessActions, setHarnessActions] = useState<Record<string, MetroraHarnessActionEvent>>({})
   const [harnessActionBusyId, setHarnessActionBusyId] = useState<string | null>(null)
@@ -423,125 +338,11 @@ export function Advisor({
     setError(null)
     setNotice(null)
   }
-  const activateHosted = () => {
-    invalidateAdvisorRequest()
-    setRuntimeChoice('hosted')
-    setHostedConsent(false)
-    void checkHostedRuntime()
-  }
-  const activateLocal = () => {
-    invalidateAdvisorRequest()
-    setRuntimeChoice(runtimeId)
-    setHostedConsent(false)
-  }
-  const updateHostedConsent = (consent: boolean) => {
-    invalidateAdvisorRequest()
-    setHostedConsent(consent)
-  }
-  const updateReasoningEffort = (effort: AdvisorReasoningEffort) => {
-    if (!reasoningEfforts.includes(effort)) return
-    invalidateAdvisorRequest()
-    setReasoningEffort(effort)
-  }
-  const updateHostedProvider = (next: AdvisorHostedProviderId) => {
-    invalidateAdvisorRequest()
-    hostedOperationGuardRef.current.setProvider(next)
-    hostedProbeController.current?.abort()
-    setHostedProvider(next)
-    setHostedModel(null)
-    setHostedConsent(false)
-    setCredentialSaving(false)
-    void checkHostedRuntime(next, true)
-  }
-  const updateHostedModel = (model: string) => {
-    invalidateAdvisorRequest()
-    setHostedModel(model)
-    setHostedConsent(false)
-  }
-  const updateLocalRuntime = (next: AdvisorLocalRuntimeId) => {
-    invalidateAdvisorRequest()
-    setRuntimeId(next)
-    setRuntimeModel(null)
-    setHostedConsent(false)
-    void checkLocalRuntime(next)
-  }
-  const updateLocalModel = (model: string) => {
-    invalidateAdvisorRequest()
-    setLocalModel(model)
-  }
-  const saveHostedCredential = async () => {
-    if (!credentialEntry.trim() || credentialSaving) return
-    const requestedProvider = hostedProvider
-    const operationId = hostedOperationGuardRef.current.startCredential(requestedProvider)
-    if (operationId === null) return
-    setCredentialSaving(true)
-    try {
-      const status = await metrora.advisorCredentialSet(requestedProvider, credentialEntry)
-      if (!hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) return
-      invalidateAdvisorRequest()
-      setNotice(status.state === 'ready' ? 'Provider credential saved in protected local storage.' : 'Provider credential was not saved: ' + status.state + '.')
-      await checkHostedRuntime(requestedProvider)
-    } catch {
-      if (hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) setNotice('Provider credential could not be saved. Enter it again.')
-    } finally {
-      if (hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) {
-        setCredentialEntry('')
-        setCredentialSaving(false)
-      }
-    }
-  }
-  const clearHostedCredential = async () => {
-    const requestedProvider = hostedProvider
-    const operationId = hostedOperationGuardRef.current.startCredential(requestedProvider)
-    if (operationId === null) return
-    try {
-      await metrora.advisorCredentialClear(requestedProvider)
-      if (!hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) return
-      invalidateAdvisorRequest()
-      setHostedConsent(false)
-      setNotice('Provider credential removed from this device.')
-      await checkHostedRuntime(requestedProvider)
-    } catch {
-      if (hostedOperationGuardRef.current.isCurrentCredential(requestedProvider, operationId)) setNotice('Provider credential could not be removed.')
-    }
-  }
   const normalizedHistoryQuery = historyQuery.trim().toLowerCase()
   const filteredConversations = conversations.filter(conversation => !normalizedHistoryQuery || [
     conversation.title,
     ...conversation.messages.map(message => message.text ?? message.answer?.conclusion ?? ''),
   ].some(value => value.toLowerCase().includes(normalizedHistoryQuery)))
-  const runtimeControls = {
-    runtimeChoice,
-    runtimeId,
-    runtimeModel,
-    llamaServerPort,
-    runtimeState,
-    hostedProvider,
-    hostedModel,
-    hostedProbe,
-    hostedConsent,
-    reasoningEffort: effectiveReasoningEffort,
-    reasoningEfforts,
-    hasHostedRuntime: Boolean(hostedRuntime),
-    configureOpen,
-    credentialEntry,
-    credentialSaving,
-    onToggleConfigure: () => setConfigureOpen(current => !current),
-    onCheckHostedRuntime: () => void checkHostedRuntime(),
-    onActivateLocal: activateLocal,
-    onHostedProviderChange: updateHostedProvider,
-    onHostedModelChange: updateHostedModel,
-    onHostedConsentChange: updateHostedConsent,
-    onReasoningEffortChange: updateReasoningEffort,
-    onCredentialEntryChange: setCredentialEntry,
-    onSaveHostedCredential: () => void saveHostedCredential(),
-    onClearHostedCredential: () => void clearHostedCredential(),
-    onCheckLocalRuntime: () => void checkLocalRuntime(),
-    onLlamaServerPortChange: setLlamaServerPort,
-    onActivateHosted: activateHosted,
-    onLocalRuntimeChange: updateLocalRuntime,
-    onLocalModelChange: updateLocalModel,
-  }
   const retryFailedRequest = () => {
     if (failedRequest) {
       setActiveConversationId(failedRequest.conversationId)
